+++ client/src/app/api/posts/[id]/route.ts
@@ -4,6 +4,13 @@ import { NextResponse } from 'next/server';
 
 export async function GET (request: Request, { params }: { params: { id: string } }): Promise<NextResponse> {
   const jwtPayload = await getJWTPayload();
+
+  if (!jwtPayload?.sub) {
+    return NextResponse.json({ error: 'Unauthenticated' }, { status: 403 });
+  }
+
+  const { id } = JSON.parse(jwtPayload.sub);
+
   const client = getClient();
   await client.connect();
 
@@ -11,7 +18,7 @@ export async function GET (request: Request, { params }: { params: { id: string 
     `select p.*, u.username, u.avatar, u.is_admin from posts p 
     inner join users u on p.user_id = u.id 
     where p.user_id = $1 and p.id = $2`,
-    [jwtPayload?.sub, params.id]
+    [id, params.id]
   );
 
   if (!foundPostRes.rowCount) {
@@ -28,12 +35,19 @@ export async function GET (request: Request, { params }: { params: { id: string 
 
 export async function DELETE (request: Request, { params }: { params: { id: string } }): Promise<NextResponse> {
   const jwtPayload = await getJWTPayload();
+
+  if (!jwtPayload?.sub) {
+    return NextResponse.json({ error: 'Unauthenticated' }, { status: 403 });
+  }
+
+  const { id } = JSON.parse(jwtPayload.sub);
+
   const client = getClient();
   await client.connect();
 
   const deletedPostRes = await client.query(
     'delete from posts where user_id = $1 and id = $2 returning *',
-    [jwtPayload?.sub, params.id]
+    [id, params.id]
   );
 
   await client.end();
@@ -51,12 +65,18 @@ export async function PATCH (request: Request, { params }: { params: { id: strin
 
   const jwtPayload = await getJWTPayload();
 
+  if (!jwtPayload?.sub) {
+    return NextResponse.json({ error: 'Unauthenticated' }, { status: 403 });
+  }
+
+  const { id } = JSON.parse(jwtPayload.sub);
+
   const client = getClient();
   await client.connect();
 
   const foundPostRes = await client.query(
     'select id from posts where user_id = $1 and id = $2',
-    [jwtPayload?.sub, params.id]
+    [id, params.id]
   );
 
   if (!foundPostRes.rowCount) {
@@ -66,7 +86,7 @@ export async function PATCH (request: Request, { params }: { params: { id: strin
 
   await client.query(
     'update posts set content = $1, updated_at = now() where user_id = $2 and id = $3',
-    [content, jwtPayload?.sub, params.id]
+    [content, id, params.id]
   );
 
   await client.end();
